Revert to 2023.3.2

+++ distributed/http/scheduler/tests/test_scheduler_http.py
@@ -503,6 +503,6 @@ async def test_check_idle(c, s, a, b):
             response = json.loads(await resp.text())
             assert isinstance(response["idle"], bool)
             assert (
-                isinstance(response["idle_since"], float)
+                isinstance(response["idle_since"], int)
                 or response["idle_since"] is None
             )
